hearing_project/requirements.txt
asgiref==3.8.1
audioread==3.0.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
decorator==5.2.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
ffmpeg-python==0.2.0
future==1.0.0
idna==3.10
joblib==1.4.2
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
msgpack==1.1.0
numba==0.61.2
numpy==2.2.5
packaging==25.0
platformdirs==4.3.7
pooch==1.8.2
pycparser==2.22
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
soundfile==0.13.1
soxr==0.5.0.post1
sqlparse==0.5.3
threadpoolctl==3.6.0
typing_extensions==4.13.2
urllib3==2.4.0
